fix: circular import; extract REFERENCE to own namespace
