+++ app/[locale]/collectibles/page-jsonld.tsx
@@ -9,7 +9,8 @@ import { normalizeUrlForJsonLd } from "@/lib/utils/url"
 import { COLLECTIBLES_BASE_URL } from "./constants"
 import type { Badge, Stats } from "./types"
 
-import { BASE_GRAPH_NODES, REFERENCE } from "@/lib/jsonld/constants"
+import { BASE_GRAPH_NODES } from "@/lib/jsonld/constants"
+import { REFERENCE } from "@/lib/jsonld/references"
 
 export default async function CollectiblesJsonLD({
   locale,
