+++ app/[locale]/community/events/page-jsonld.tsx
@@ -10,7 +10,8 @@ import { normalizeUrlForJsonLd } from "@/lib/utils/url"
 import { communityHubSchemas } from "@/data/community-hub-schemas"
 import communityHubs from "@/data/community-hubs"
 
-import { BASE_GRAPH_NODES, REFERENCE } from "@/lib/jsonld/constants"
+import { BASE_GRAPH_NODES } from "@/lib/jsonld/constants"
+import { REFERENCE } from "@/lib/jsonld/references"
 
 function buildHubSchemaNodes(
   hub: (typeof communityHubs)[number],
