+++ app/[locale]/videos/[slug]/page-jsonld.tsx
@@ -8,7 +8,8 @@ import { stripMarkdown } from "@/lib/utils/md"
 import { toIsoDuration } from "@/lib/utils/time"
 import { normalizeUrlForJsonLd } from "@/lib/utils/url"
 
-import { BASE_GRAPH_NODES, REFERENCE } from "@/lib/jsonld/constants"
+import { BASE_GRAPH_NODES } from "@/lib/jsonld/constants"
+import { REFERENCE } from "@/lib/jsonld/references"
 import { resolveAuthorsFromFrontmatter } from "@/lib/jsonld/utils"
 
 export default async function VideoPageJsonLD({
