+++ src/data/placeholders/content-developers-tutorials-gasless-token-data.json
@@ -1,0 +1,14 @@
+{
+  "/content/developers/tutorials/gasless-token/fig-1-opaque-calldata.png": {
+    "hash": "50bd0276",
+    "base64": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAANCAYAAACgu+4kAAAACXBIWXMAAA7DAAAOwwHHb6hkAAABxElEQVR4nK1SW0sbQRgdwSt4RUR9KIhWUmoQfBCF+iYI/QXWhyj+A+k/EHozUbxE3MR4pUolGxUkBB/MWuIGREutLphN20QkKlQTlPUlJOqR+TYV64sXeuAwwzfnnPmYb5iiqNC0c0SjMcRip9A0DZvfFHw0O+AYE9E/OA1L/yQCaginZ2ek4+Se9Y0tsFA4gmQigds4OPyDmS9uLC5JxDnXMoXfxfZOgAfsQZZl2O0OrHgljE9MQbDZMSKMwGq1wmYT4Pa4MT4xicGhYTidIj6ZLaTlXTHeQVfXW+TmF5NAFOdpff+hBz2WXlqd4gI+z8xCsI3C5ZpHX98A/H6/HhA5OEJHRyfKyp+hrc2EGmMd6htewfDCSGxobEJFZTWqDTWoem4gcm3rGxN+hfbBNja3SJiVXQDGMsHSc1FUXEodFRSWkDktI4/IWE5Klw5Te6cesBsI4qWxFowxCuDC5pbXkKRV4orXC0n6Cp8sw7e2dlMPh/eg7P7U34AXurvfweNZhqoGEY/HifeBphBQf9Ps/+K2OZlI4vLqKrXXR31xeXFznhpj5B/jQ26+8w/0gKfg/wR8/6HQHz8+iT6K3OOT13ENdsCgCOj72ZoAAAAASUVORK5CYII="
+  },
+  "/content/developers/tutorials/gasless-token/fig-2-transparent-signature.png": {
+    "hash": "924f4956",
+    "base64": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAOCAYAAAAmL5yKAAAACXBIWXMAAA7DAAAOwwHHb6hkAAABkUlEQVR4nK1QPUiCURR9BQ1NTUJzDWG/Rj/Sz1BDGBREaGFGBQ39CC5RQWXQEFG0RUs0SEFb2NRUNkRRWET5gZoa6Yf6qVEpDdaQJ979cDAJMjpw3r2cd9/hvMvsV7d49It4eBSpcjrdPthOLnB6ZseR7ZzINT6TZjAs4dJ+AxaNxvAdV9cCDCNTMJqWMDo2B53eBIdwnzUnRZ7APB4fAgER8XiCaliSiJFIBH6/H6FwCK+vL3SX1uT5OEQxKBsYBodQUlqGjY1NTM/Mor29Az09WrS0thEnJ00wDA5D09lFva5PD01nNwTBCcbd+vsHwBiD3jCEefMitra2MT5hxOraOpLJJMXlNd3LSMkJEok3zC+YyZ3lF6KgsAhqdTMZMpaHGlU9GhqbUKOqg7K8GhVVtaTxlC63l3/BC6WykmIdH9uwvLxCCX4DWuL7+wcslh0oFMXo1fZRb7UegOuZkX8wuL1zwOW+pw1z5gIyiMWeM8TsZeVogFwT8IPjM5XKiRwhKfYPCXb39mE9OPwT+dsvg5z1eJ2JS34AAAAASUVORK5CYII="
+  },
+  "/content/developers/tutorials/gasless-token/fig-3-truncated-addresses.png": {
+    "hash": "8973a2b6",
+    "base64": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAoAAAAQCAYAAAAvf+5AAAAACXBIWXMAAA7DAAAOwwHHb6hkAAAB0ElEQVR4nF2R3WoTURDHBwSv1VAFEYNe1Auv/Eh9A++EgtA+iX0Ir/oY3gUilBYhhIrQJGa7m5p0zX402XQ92+TsdrskbtgQ/zIT8MIDw5lzzm9m/meG3t16ire3n2OLHuAVlfCC7oi9pBIqdB9v6CFe0wbo0/5HfD2oYZJcI45j/L+WyyVc2wHZ7SbCCx8qihCGIbTW0HGMKIqQpqnAKlKgPwBmsxkODg9R+1zD0dEX1Ot1VKtVnJqmgEEQgBaLhZTkaLYkSXCdpsiyTHZ+50rE0NnZDzkwyOdYTxHHCfr9c1hWF57ngzgDg7b9E543hG+bcHpNuTs5aWIwcNYg6wuCMbSORWua3uAmyzCf5eLP578xHI5ADLQ7hmRQKkKsE2idSJZutwfHceH5/ho8NS15iNQU6irAuWvg+PgbGo0G2u3vcBxnDRqGiV6vjyC4xGR6hYuRi9EwEMi0unA9D8TNZY2TyVR0/goVDMNCq9USsNMx4DHIAP+OBXNLZDJaiz8eX4r2wWDA7Unh+4H8kH3uY1EsxfJ8gdVqJeOkPM+lpGlaUEqhKApsbj4DEeHuvZLsOzu7IJ4GR/i+/28629vvUS4/QaWyhUePy/iwt4e/C3sb8fdXuNsAAAAASUVORK5CYII="
+  }
+}
