+++ src/lib/jsonld/references.ts
@@ -1,0 +1,14 @@
+export const REFERENCE = {
+  /**
+   * Reference to Ethereum Foundation organization (for use when the full object is already defined elsewhere)
+   */
+  ETHEREUM_FOUNDATION: { "@id": "https://ethereum.foundation/#organization" },
+  /**
+   * Reference to Ethereum Community organization (for use when the full object is already defined elsewhere)
+   */
+  ETHEREUM_COMMUNITY: { "@id": "https://ethereum.org/#community-organization" },
+  /**
+   * Reference to the ethereum.org WebSite entity (for use when the full object is already defined elsewhere in the graph)
+   */
+  ETHEREUM_ORG_WEBSITE: { "@id": "https://ethereum.org/#website" },
+} as const
